Migrate pyrat-headless to Match and accept_players
Slice 6 of the host restructure: the headless binary now drives the
new lifecycle. main.rs goes from 453 to 271 lines as the manual
plumbing (accept_connections, run_setup, run_playing, HostCommand
shutdown drain) collapses into accept_players + Match::new + match.run().
The JSON record builder moves out to record.rs.

accept_players consumes the Identify frame as part of the multi-slot
handshake, so Match never sees a BotIdentified protocol event. Headless
emits MatchEvent::BotIdentified itself from each TcpPlayer.identity()
right after accept_players returns — same channel, same consumer, so
the JSON record's players[] field and the tracing span fields keep
working. CLI: --think-margin is dropped (it was a silent no-op against
the new PlayingConfig); --configure-timeout-ms is added (default 5000)
to feed SetupTiming.configure_timeout independently of preprocessing
and overall accept timeouts.

test_bot.rs is rewritten as a 31-line SDK-based random bot. The new
wire's hash-verified handshake means a hand-rolled subprocess test bot
would have to compute Zobrist hashes off MatchConfig anyway; using the
SDK keeps the test honest — anything that breaks the SDK breaks this
test, and vice versa. flatbuffers and the legacy wire imports drop
from headless's Cargo.toml; pyrat-sdk is added.

This is the first live e2e gate. tests/e2e.rs runs the headless binary
against two test_bot subprocesses and asserts a valid game record JSON.
The full new stack — codec, framing, accept_players, TcpPlayer, Match,
JSON output — is exercised end-to-end for the first time.

Tests: tests/e2e.rs:headless_with_test_bots passes; 138 host tests
unchanged.
