+++ server/headless/Cargo.toml
@@ -14,9 +14,9 @@ path = "src/bin/test_bot.rs"
 
 [dependencies]
 clap = { version = "4", features = ["derive"] }
-flatbuffers = "25.2.10"
 pyrat-host = { path = "../host" }
 pyrat-rust = { path = "../../engine", default-features = false }
+pyrat-sdk = { path = "../../sdk/rust" }
 rand = "0.10"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
